+++ pkg/slayers/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "extn.go",
         "l4.go",
         "layertypes.go",
+        "pkt_auth.go",
         "scion.go",
         "scmp.go",
         "scmp_msg.go",
@@ -33,6 +34,7 @@ go_test(
         "bfd_test.go",
         "export_test.go",
         "extn_test.go",
+        "pkt_auth_test.go",
         "scion_test.go",
         "scmp_msg_test.go",
         "scmp_test.go",
