+++ pkg/slayers/extn.go
@@ -15,7 +15,6 @@
 package slayers
 
 import (
-	"encoding/binary"
 	"fmt"
 
 	"github.com/google/gopacket"
@@ -445,203 +444,4 @@ func (s *EndToEndExtnSkipper) CanDecode() gopacket.LayerClass {
 
 func (e *EndToEndExtnSkipper) NextLayerType() gopacket.LayerType {
 	return scionNextLayerTypeAfterE2E(e.NextHdr)
-}
-
-// PacketAuthSPI is the identifier for the key used for the
-// packet authentication option. DRKey values are in the
-// range [1, 2^21-1].
-type PacketAuthSPI uint32
-
-type PacketAuthDRKeyType uint8
-
-const (
-	PacketAuthASHost PacketAuthDRKeyType = iota
-	PacketAuthHostHost
-)
-
-type PacketAuthDRKeyDirection uint8
-
-const (
-	PacketAuthSenderSide PacketAuthDRKeyDirection = iota
-	PacketAuthReceiverSide
-)
-
-type PacketAuthDRKeyEpochType uint8
-
-const (
-	PacketAuthLater PacketAuthDRKeyEpochType = iota
-	PacketAuthEarlier
-)
-
-func (p PacketAuthSPI) Type() PacketAuthDRKeyType {
-	if p&(1<<18) == 0 {
-		return PacketAuthASHost
-	}
-	return PacketAuthHostHost
-}
-
-func (p PacketAuthSPI) Direction() PacketAuthDRKeyDirection {
-	if p&(1<<17) == 0 {
-		return PacketAuthSenderSide
-	}
-	return PacketAuthReceiverSide
-}
-
-func (p PacketAuthSPI) Epoch() PacketAuthDRKeyEpochType {
-	if p&(1<<16) == 0 {
-		return PacketAuthLater
-	}
-	return PacketAuthEarlier
-}
-
-func (p PacketAuthSPI) DRKeyProto() uint16 {
-	return uint16(p)
-}
-
-func (p PacketAuthSPI) IsDRKey() bool {
-	return p > 0 && p < (1<<21)
-}
-
-func MakePacketAuthSPIDRKey(
-	proto uint16,
-	drkeyType PacketAuthDRKeyType,
-	dir PacketAuthDRKeyDirection,
-	epoch PacketAuthDRKeyEpochType,
-) PacketAuthSPI {
-
-	if proto < 1 {
-		panic("Invalid proto identifier value")
-	}
-	if drkeyType > 1 {
-		panic("Invalid DRKeyType value")
-	}
-	if dir > 1 {
-		panic("Invalid DRKeyDirection value")
-	}
-	if epoch > 1 {
-		panic("Invalid DRKeyEpochType value")
-	}
-	spi := uint32((drkeyType & 0x1)) << 18
-	spi |= uint32((dir & 0x1)) << 17
-	spi |= uint32((epoch & 0x1)) << 16
-	spi |= uint32(proto)
-
-	return PacketAuthSPI(spi)
-}
-
-// PacketAuthAlg is the enumerator for authenticator algorithm types in the
-// packet authenticator option.
-type PacketAuthAlg uint8
-
-const (
-	PacketAuthCMAC PacketAuthAlg = iota
-	PacketAuthSHA1_AES_CBC
-)
-
-// PacketAuthOption wraps an EndToEndOption of OptTypeAuthenticator.
-// This can be used to serialize and parse the internal structure of the packet authenticator
-// option.
-type PacketAuthOption struct {
-	*EndToEndOption
-}
-
-// NewPacketAuthOption creates a new EndToEndOption of
-// OptTypeAuthenticator, initialized with the given SPAO data.
-func NewPacketAuthOption(
-	spi PacketAuthSPI,
-	alg PacketAuthAlg,
-	ts uint32,
-	sn uint32,
-	auth []byte,
-) PacketAuthOption {
-
-	o := PacketAuthOption{EndToEndOption: new(EndToEndOption)}
-	o.Reset(spi, alg, ts, sn, auth)
-	return o
-}
-
-// ParsePacketAuthOption parses o as a packet authenticator option.
-// Performs minimal checks to ensure that SPI, algorithm, timestamp, RSV, and
-// sequence number are set.
-// Checking the size and content of the Authenticator data must be done by the
-// caller.
-func ParsePacketAuthOption(o *EndToEndOption) (PacketAuthOption, error) {
-	if o.OptType != OptTypeAuthenticator {
-		return PacketAuthOption{},
-			serrors.New("wrong option type", "expected", OptTypeAuthenticator, "actual", o.OptType)
-	}
-	if len(o.OptData) < 12 {
-		return PacketAuthOption{},
-			serrors.New("buffer too short", "expected", 12, "actual", len(o.OptData))
-	}
-	return PacketAuthOption{o}, nil
-}
-
-// Reset reinitializes the underlying EndToEndOption with the SPAO data.
-// Reuses the OptData buffer if it is of sufficient capacity.
-func (o PacketAuthOption) Reset(
-	spi PacketAuthSPI,
-	alg PacketAuthAlg,
-	ts uint32,
-	sn uint32,
-	auth []byte,
-) {
-
-	if ts >= (1 << 24) {
-		panic("Timestamp value should be smaller than 2^24")
-	}
-	if sn >= (1 << 24) {
-		panic("Sequence number should be smaller than 2^24")
-	}
-
-	o.OptType = OptTypeAuthenticator
-
-	n := 12 + len(auth) // 4 + 1 + 3 + 1 + 3 + len(data)
-	if n <= cap(o.OptData) {
-		o.OptData = o.OptData[:n]
-	} else {
-		o.OptData = make([]byte, n)
-	}
-	binary.BigEndian.PutUint32(o.OptData[:4], uint32(spi))
-	o.OptData[4] = byte(alg)
-	o.OptData[5] = byte(ts >> 16)
-	o.OptData[6] = byte(ts >> 8)
-	o.OptData[7] = byte(ts)
-	o.OptData[8] = byte(0)
-	o.OptData[9] = byte(sn >> 16)
-	o.OptData[10] = byte(sn >> 8)
-	o.OptData[11] = byte(sn)
-	copy(o.OptData[12:], auth)
-
-	o.OptAlign = [2]uint8{4, 2}
-	// reset unused/implicit fields
-	o.OptDataLen = 0
-	o.ActualLength = 0
-}
-
-// SPI returns returns the value set in the homonym field in the extension.
-func (o PacketAuthOption) SPI() PacketAuthSPI {
-	return PacketAuthSPI(binary.BigEndian.Uint32(o.OptData[:4]))
-}
-
-// Algorithm returns the algorithm type stored in the data buffer.
-func (o PacketAuthOption) Algorithm() PacketAuthAlg {
-	return PacketAuthAlg(o.OptData[4])
-}
-
-// Timestamp returns the value set in the homonym field in the extension.
-func (o PacketAuthOption) Timestamp() uint32 {
-	return uint32(o.OptData[5])<<16 + uint32(o.OptData[6])<<8 + uint32(o.OptData[7])
-}
-
-// SequenceNumber returns the value set in the homonym field in the extension.
-func (o PacketAuthOption) SequenceNumber() uint32 {
-	return uint32(o.OptData[9])<<16 + uint32(o.OptData[10])<<8 + uint32(o.OptData[11])
-}
-
-// Authenticator returns slice of the underlying auth buffer.
-// Changes to this slice will be reflected on the wire when
-// the extension is serialized.
-func (o PacketAuthOption) Authenticator() []byte {
-	return o.OptData[12:]
 }
