+++ pkg/slayers/extn_test.go
@@ -15,7 +15,6 @@
 package slayers_test
 
 import (
-	"encoding/binary"
 	"fmt"
 	"testing"
 
@@ -541,158 +540,4 @@ func prepRawPacketWithExtn(t *testing.T, extns ...slayers.L4ProtocolType) []byte
 	buf.AppendBytes(9) // dummy UDP with 1 byte payload
 
 	return buf.Bytes()
-}
-
-var spi = slayers.MakePacketAuthSPIDRKey(
-	1,
-	slayers.PacketAuthASHost,
-	slayers.PacketAuthReceiverSide,
-	slayers.PacketAuthLater)
-var algo = slayers.PacketAuthSHA1_AES_CBC
-var ts = binary.LittleEndian.Uint32([]byte{1, 2, 3, 0})
-var sn = binary.LittleEndian.Uint32([]byte{4, 5, 6, 0})
-var optAuthMAC = []byte("16byte_mac_foooo")
-
-// +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-// |   NextHdr=UDP |     ExtLen    |  OptType=2    |  OptDataLen   |
-// +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-// |                   Security Parameter Index                    |
-// +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-// |    Algorithm  |                    Timestamp                  |
-// +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-// |      RSV      |                  Sequence Number              |
-// +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-// |                                                               |
-// +                                                               +
-// |                                                               |
-// +                        16-octet MAC data                      +
-// |                                                               |
-// +                                                               +
-// |                                                               |
-// +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-
-var rawE2EOptAuth = append(
-	[]byte{
-		0x11, 0x7, 0x2, 0x1c,
-		0x0, 0x2, 0x0, 0x1,
-		0x1, 0x3, 0x2, 0x1,
-		0x0, 0x6, 0x5, 0x4,
-	},
-	optAuthMAC...,
-)
-
-func TestOptAuthenticatorSerialize(t *testing.T) {
-	cases := []struct {
-		name    string
-		spi     slayers.PacketAuthSPI
-		algo    slayers.PacketAuthAlg
-		ts      uint32
-		sn      uint32
-		optAuth []byte
-		panics  bool
-	}{
-		{
-			name:    "correct",
-			spi:     spi,
-			algo:    algo,
-			ts:      ts,
-			sn:      sn,
-			optAuth: optAuthMAC,
-			panics:  false,
-		},
-		{
-			name:    "bad_ts",
-			spi:     spi,
-			algo:    algo,
-			ts:      binary.LittleEndian.Uint32([]byte{0, 0, 0, 1}),
-			sn:      sn,
-			optAuth: optAuthMAC,
-			panics:  true,
-		},
-		{
-			name:    "bad_sn",
-			spi:     spi,
-			algo:    algo,
-			ts:      ts,
-			sn:      binary.LittleEndian.Uint32([]byte{0, 0, 0, 1}),
-			optAuth: optAuthMAC,
-			panics:  true,
-		},
-	}
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			if c.panics {
-				assert.Panics(t, func() {
-					slayers.NewPacketAuthOption(c.spi, c.algo, c.ts, c.sn, c.optAuth)
-				},
-					"The code did not panic",
-				)
-				return
-			}
-
-			spao := slayers.NewPacketAuthOption(c.spi, c.algo, c.ts, c.sn, c.optAuth)
-
-			e2e := slayers.EndToEndExtn{}
-			e2e.NextHdr = slayers.L4UDP
-			e2e.Options = []*slayers.EndToEndOption{spao.EndToEndOption}
-
-			b := gopacket.NewSerializeBuffer()
-			opts := gopacket.SerializeOptions{FixLengths: true}
-			assert.NoError(t, e2e.SerializeTo(b, opts), "SerializeTo")
-			assert.Equal(t, rawE2EOptAuth, b.Bytes(), "Raw Buffer")
-		})
-	}
-}
-
-func TestOptAuthenticatorDeserialize(t *testing.T) {
-	e2e := slayers.EndToEndExtn{}
-
-	_, err := e2e.FindOption(slayers.OptTypeAuthenticator)
-	assert.Error(t, err)
-
-	assert.NoError(t, e2e.DecodeFromBytes(rawE2EOptAuth, gopacket.NilDecodeFeedback))
-	assert.Equal(t, slayers.L4UDP, e2e.NextHdr, "NextHeader")
-	optAuth, err := e2e.FindOption(slayers.OptTypeAuthenticator)
-	require.NoError(t, err, "FindOption")
-	auth, err := slayers.ParsePacketAuthOption(optAuth)
-	require.NoError(t, err, "ParsePacketAuthOption")
-	assert.Equal(t, spi, auth.SPI(), "SPI")
-	assert.Equal(t, slayers.PacketAuthASHost, auth.SPI().Type())
-	assert.Equal(t, slayers.PacketAuthReceiverSide, auth.SPI().Direction())
-	assert.Equal(t, slayers.PacketAuthLater, auth.SPI().Epoch())
-	assert.Equal(t, true, auth.SPI().IsDRKey())
-	assert.Equal(t, algo, auth.Algorithm(), "Algorithm Type")
-	assert.Equal(t, ts, auth.Timestamp(), "Timestamp")
-	assert.Equal(t, sn, auth.SequenceNumber(), "Sequence Number")
-	assert.Equal(t, optAuthMAC, auth.Authenticator(), "Authenticator data (MAC)")
-}
-
-func TestMakePacketAuthSPIDrkey(t *testing.T) {
-	spi := slayers.MakePacketAuthSPIDRKey(
-		1,
-		slayers.PacketAuthASHost,
-		slayers.PacketAuthSenderSide,
-		slayers.PacketAuthLater,
-	)
-	assert.EqualValues(t, binary.BigEndian.Uint32([]byte{0, 0, 0, 1}), spi)
-}
-
-func TestOptAuthenticatorDeserializeCorrupt(t *testing.T) {
-	optAuthCorrupt := slayers.EndToEndOption{
-		OptType: slayers.OptTypeAuthenticator,
-		OptData: []byte{},
-	}
-	e2e := slayers.EndToEndExtn{}
-	e2e.NextHdr = slayers.L4UDP
-	e2e.Options = []*slayers.EndToEndOption{&optAuthCorrupt}
-
-	b := gopacket.NewSerializeBuffer()
-	opts := gopacket.SerializeOptions{FixLengths: true}
-	assert.NoError(t, e2e.SerializeTo(b, opts), "SerializeTo")
-
-	assert.NoError(t, e2e.DecodeFromBytes(b.Bytes(), gopacket.NilDecodeFeedback))
-	optAuth, err := e2e.FindOption(slayers.OptTypeAuthenticator)
-	require.NoError(t, err, "FindOption")
-	_, err = slayers.ParsePacketAuthOption(optAuth)
-	require.Error(t, err, "ParsePacketAuthOption should fail")
 }
